fix existing unit tests
